Introduce useApi hook
