+++ src/hooks/apis/useCloudAccounts.js
@@ -1,14 +1,14 @@
 import { useQuery } from 'react-query';
-
-import apiRequest from '../../services/apiRequest';
+import useApi from './useApi';
 
 export const queryKey = 'cloud_accounts';
 
-const getCloudAccounts = async () => {
+const getCloudAccounts = async (apiRequest) => {
   const { data } = await apiRequest.get(`/api/rhacs/v1/cloud_accounts`);
   return data;
 };
 
 export default function useCloudAccounts() {
-  return useQuery([queryKey], () => getCloudAccounts());
+  const apiRequest = useApi();
+  return useQuery([queryKey], () => getCloudAccounts(apiRequest));
 }
