+++ src/hooks/apis/useCloudRegions.js
@@ -1,12 +1,13 @@
 import { useQuery } from 'react-query';
 
-import apiRequest from '../../services/apiRequest';
 import { getQueryString } from '../../utils/queryString';
+import useApi from './useApi';
 
 export const queryKey = (query) => ['cloud_regions', query];
 
 const getCloudRegions = async ({ provider, instanceType }) => {
   const queryString = getQueryString({ instance_type: instanceType });
+  const apiRequest = useApi();
   const { data } = await apiRequest.get(
     `/api/rhacs/v1/cloud_providers/${provider}/regions?${queryString}`
   );
