+++ src/hooks/apis/useInstances.js
@@ -1,9 +1,10 @@
 import { useQuery } from 'react-query';
 
-import apiRequest from '../../services/apiRequest';
 import { getQueryString } from '../../utils/queryString';
+import useApi from './useApi';
 
 const getInstances = async ({ query }) => {
+  const apiRequest = useApi();
   const queryString = getQueryString(query);
   const { data } = await apiRequest.get(
     `/api/rhacs/v1/centrals?${queryString}`
